update edit apartment

+++ CMSDiamondStay/Models/DashboardAdmin.cs
@@ -8,8 +8,8 @@ namespace CMSDiamondStay.Models
     public class DashboardAdmin
     {
         public int total_user { get; set; }
-        public decimal total { get; set; }
-        public decimal total_month { get; set; }
+        public double total { get; set; }
+        public double total_month { get; set; }
         public int total_apartment { get; set; }
         public int total_booking { get; set; }
         public int total_booking_finish { get; set; }
